Move the installation of git to a separate function
git would likely be installed on a new macOS system, if it is not, we will install it using brew.

Fixes the following error on GitHub macOS runners:
```
Error: The `brew link` step did not complete successfully
The formula built, but is not symlinked into /usr/local
Could not symlink etc/bash_completion.d/git-completion.bash
Target /usr/local/etc/bash_completion.d/git-completion.bash
is a symlink belonging to git@2.35.1. You can unlink it:
  brew unlink git@2.35.1

To force the link and overwrite all conflicting files:
  brew link --overwrite git

To list all files that would be deleted:
  brew link --overwrite --dry-run git
```